+++ packages/expo-router/ios/LinkPreview/LinkPreviewNativeActionView.swift
@@ -11,6 +11,7 @@ class LinkPreviewNativeActionView: RouterViewWithLogger, LinkPreviewMenuUpdatabl
       updateMenu()
     }
   }
+  @NativeActionProp(updateAction: true, updateMenu: true) var imageRenderingMode: ImageRenderingMode?
   @NativeActionProp(updateAction: true, updateMenu: true) var destructive: Bool?
   @NativeActionProp(updateAction: true, updateMenu: true) var disabled: Bool = false
 
@@ -56,7 +57,8 @@ class LinkPreviewNativeActionView: RouterViewWithLogger, LinkPreviewMenuUpdatabl
 
   private var image: UIImage? {
     if let customImage = customImage {
-      return customImage.ref
+      let renderingMode: UIImage.RenderingMode = imageRenderingMode == .template ? .alwaysTemplate : .alwaysOriginal
+      return customImage.ref.withRenderingMode(renderingMode)
     }
     if let icon = icon {
       return UIImage(systemName: icon)